+++ scripts/git/git.js
@@ -1,4 +1,5 @@
-const typeLabel = "type=";
+const { getArgValue, isListEmpty, firstElement } = require("./../utils/utils");
+const typeLabel = "type";
 
 const execute = (args) => {
     const argsObj = buildArgsObj(args);
@@ -7,20 +8,9 @@ const execute = (args) => {
 
 const buildArgsObj = (argsObj) => {
     return {
-        type: getType(argsObj),
+        type: getArgValue(argsObj, typeLabel),
         args: argsObj
     };
-}
-
-const getType = (argsObj) => {
-    const typeList = argsObj.filter(arg => arg.includes(typeLabel));
-
-    if (typeList.length < 1) {
-        console.error('Argument type is missing!');
-        process.exit(1);
-    }
-
-    return typeList[0].replace(typeLabel, "");
 }
 
 const executeScript = (argsObj) => {
@@ -34,12 +24,12 @@ const getSelectedType = (argsObj) => {
     const scriptSettings = require('./settings.json');
     const selectedTypeList = scriptSettings.types.filter(types => types.type===argsObj.type);
 
-    if (selectedTypeList.length < 1) {
+    if (isListEmpty(selectedTypeList)) {
         console.error('Type is not defined in the settings.json.');
         process.exit(1);
     }
 
-    return selectedTypeList[0];
+    return firstElement(selectedTypeList);
 }
 
 module.exports.execute = execute;