+++ scripts/runners/operations/create.js
@@ -1,21 +1,31 @@
 const fs = require("fs");
-const createFilesLabel="createFiles=";
-let runnersPathFrom="projects/{project}/runners";
-let runnersPathTo="../{project}/.idea/runConfigurations";
+const { validateDestinationPath, validateDestinationPathNoCreation, getFiles, getArgValue, convertStringToBoolean } = require("../../utils/utils");
+const createFilesLabel = "createFiles";
+let runnersPathFrom = "projects/{project}/runners";
+let projectPath = "../{project}";
+let ideaPath = "/.idea";
+let runnersPath = "/runConfigurations";
 
 const execute = (args) => {
     console.log('Starting process for the project: ' + args.project);
     const argsObj = buildArgsObj(args);
 
-    runnersPathFrom=runnersPathFrom.replace("{project}", argsObj.project);
-    runnersPathTo=runnersPathTo.replace("{project}", argsObj.project);
-    validateDestinationPath(runnersPathTo);
+    const modifiedProjectPath = projectPath.replace("{project}", argsObj.project);
+    if(!validateDestinationPathNoCreation(modifiedProjectPath)) {
+        console.error(`Project ${argsObj.project} does not exist`);
+        return;
+    }
+
+    const modifiedRunnersPathFrom = runnersPathFrom.replace("{project}", argsObj.project);
+    runnersPath = modifiedProjectPath + ideaPath + runnersPath;
+    validateDestinationPath(modifiedProjectPath + ideaPath);
+    validateDestinationPath(runnersPath);
 
     const fileData = {
-        fromListOfFiles: getFiles(runnersPathFrom),
-        fromFolderPath: runnersPathFrom,
-        toListOfFiles: getFiles(runnersPathTo),
-        toFolderPath: runnersPathTo,
+        fromListOfFiles: getFiles(modifiedRunnersPathFrom),
+        fromFolderPath: modifiedRunnersPathFrom,
+        toListOfFiles: getFiles(runnersPath),
+        toFolderPath: runnersPath,
         createFiles: argsObj.createFiles
     }
     copyFiles(fileData);
@@ -23,28 +33,16 @@ const execute = (args) => {
 
 const buildArgsObj = (argsObj) => {
     return {
-        createFiles: getCreateFile(argsObj.args),
+        createFiles: convertStringToBoolean(getArgValue(argsObj.args, createFilesLabel)),
         project: argsObj.project
     };
 }
 
-const getCreateFile = (args) => {
-    let createFilesArgList = args.filter(arg => arg.includes(createFilesLabel));
-
-    if(createFilesArgList < 1) {
-        console.error("No createFiles setUp using the default value.");
-        return false;
-    }
-
-    const createFilesArg = createFilesArgList[0].replace(createFilesLabel, "");
-    return String(createFilesArg).toLowerCase() === "true";
-}
-
 const copyFiles = (fileData) => {
     let fileList = fileData.fromListOfFiles;
-    if(!fileData.createFiles) {
+    if (!fileData.createFiles) {
         fileData.toListOfFiles.forEach(toFile => {
-            toFile=toFile.replace(".run", "");
+            toFile = toFile.replace(".run", "");
             const index = fileList.indexOf(toFile);
             if (index > -1) {
                 fileList.splice(index, 1);
@@ -62,9 +60,5 @@ const copyFiles = (fileData) => {
     });
     console.log("Files created where: " + valueCount);
 }
-
-const getFiles = (path) => fs.readdirSync(path).filter(file => fs.statSync(path + '/' + file).isFile());
-
-const validateDestinationPath = (dir) => !fs.existsSync(dir) ? fs.mkdirSync(dir) : undefined;
 
 module.exports.execute = execute;