+++ scripts/runners/runners.js
@@ -1,6 +1,7 @@
-const typeLabel = "type=";
 const fs = require("fs");
-const projectsLabel = "projects=";
+const { firstElement, getArgValue, isListEmpty } = require("../utils/utils");
+const typeLabel = "type";
+const projectsLabel = "projects";
 const projectsPath = "./projects";
 
 const execute = (args) => {
@@ -10,31 +11,21 @@ const execute = (args) => {
 
 const buildArgsObj = (argsObj) => {
     return {
-        type: getType(argsObj),
+        type: getArgValue(argsObj, typeLabel),
         args: argsObj,
         projects: getProjectList(argsObj)
     };
 }
 
-const getType = (argsObj) => {
-    const typeList = argsObj.filter(arg => arg.includes(typeLabel));
-
-    if (typeList.length < 1) {
-        console.error('Argument type is missing!');
-        process.exit(1);
-    }
-
-    return typeList[0].replace(typeLabel, "");
-}
-
 const getProjectList = (argsObj) => {
-    const projectList = argsObj.filter(arg => arg.includes(projectsLabel));
+    const projectList = getArgValue(argsObj, projectsLabel);
 
-    if (projectList.length >= 1) {
-        return projectList[0].replace(projectsLabel, "").split(",");
+    console.log(projectList);
+    if (projectList === undefined) {
+        return getDirectories();
     }
 
-    return getDirectories();
+    return projectList.split(",");
 }
 
 const getDirectories = () => fs.readdirSync(projectsPath).filter(file => fs.statSync(projectsPath + '/' + file).isDirectory() && !file.includes("demo-"));
@@ -53,12 +44,12 @@ const getSelectedTypeObj = (argsObj) => {
     const scriptSettings = require('./settings.json');
     const selectedTypeList = scriptSettings.types.filter(types => types.type === argsObj.type);
 
-    if (selectedTypeList.length < 1) {
+    if (isListEmpty(selectedTypeList)) {
         console.error('Type is not defined in the settings.json.');
         process.exit(1);
     }
 
-    return selectedTypeList[0];
+    return firstElement(selectedTypeList);
 }
 
 module.exports.execute = execute;